Some test refactoring, and adding descriptions to IMethodResponse
