+++ src/sdkModels.spec.ts
@@ -1,12 +1,29 @@
-import { CriteriaToSet, IMethod, ITagList, SearchCriterion, SearchCriterionTerm, SetToCriteria } from './sdkModels'
-import { TestConfig } from '../typescript/looker/rtl/nodeSettings.spec'
-import { specFromFile } from './sdkGenerator'
+/*
+ * The MIT License (MIT)
+ *
+ * Copyright (c) 2019 Looker Data Sciences, Inc.
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in
+ * all copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+ * THE SOFTWARE.
+ */
 
-const config = TestConfig()
-export const apiTestModel = specFromFile(
-  `${config.testPath}openApiRef.json`,
-  `${config.testPath}swaggerRef.json`
-)
+import { CriteriaToSet, IMethod, ITagList, SearchCriterion, SearchCriterionTerm, SetToCriteria } from './sdkModels'
+import { apiTestModel } from './testUtils'
 
 describe('sdkModels', () => {
 
@@ -64,6 +81,13 @@ describe('sdkModels', () => {
       expect(method.responseIsBoth()).toEqual(true)
     })
 
+    it('each response is described', () => {
+      const method = apiTestModel.methods['run_look']
+      expect(method.responses.length).toBeGreaterThan(0)
+      method.responses.forEach((r => {
+        expect(r.description).not.toEqual("")
+      }))
+    })
   })
 
   describe('required properties', () => {
